+++ src/main/java/com/mochibit/defcon/lifecycle/Lifecycled.kt
@@ -21,7 +21,7 @@ package com.mochibit.defcon.lifecycle
 
 interface Lifecycled {
     fun start()
-    fun update(delta: Double)
+    fun update(delta: Float)
 
     fun stop()
 }
